epoch plot

+++ imdb/run_imdb.py
@@ -228,8 +228,8 @@ def train(model,
                         output = torch.sigmoid(output)
 
                         # ugly but it works
-                        y_true_val.extend(labels.cpu().numpy().tolist())
-                        y_pred_val.extend(output.cpu().numpy().tolist())
+                        y_true_val.extend(labels.cpu().numpy().astype(float).tolist())
+                        y_pred_val.extend(output.cpu().numpy().astype(float).tolist())
 
                         loss = criterion(output, labels)
                         valid_running_loss += loss.item()
